I added tomato
Imagine that you’re a horticulturalist with limited land and space to grow. And imagine you are what they call an original thinker, unusually gifted in lateral thinking. Well, one day you’re minding your own business, eating your mashed potatoes with a tomato salad on the side, and you get to think... what if you could grow potatoes and tomatoes on one plant? That would give you a double crop and you could produce twice as much food on that small plot of land that you own? If this was a cartoon, a flashing light bulb would appear above your head!

You grow a huge wild beard, and start digging up the bodies of deceased tomatoes and potatoes. You build a giant machine that uses enormous amounts of electricity and you start to laugh like a maniac as you create your own vegetable-Frankenstein.

It might sound crazy, but it is not as unrealistic as it seems. Okay, we can leave out the Frankenstein scene, and actually no genetic modification is necessary either. That’s because tomatoes and potatoes are actually related to each other. They are both members of the nightshade family just like sweet peppers, chilli peppers, eggplants, tomatillos, tamarios, pepinos, pimentos, paprika, and cayenne peppers.